+++ jaorm-core/src/main/java/io/github/ulisse1996/jaorm/spi/QueriesService.java
@@ -1,22 +1,32 @@
 package io.github.ulisse1996.jaorm.spi;
 
-import io.github.ulisse1996.jaorm.ServiceFinder;
 import io.github.ulisse1996.jaorm.BaseDao;
 import io.github.ulisse1996.jaorm.DaoImplementation;
+import io.github.ulisse1996.jaorm.ServiceFinder;
 import io.github.ulisse1996.jaorm.entity.EntityDelegate;
+import io.github.ulisse1996.jaorm.spi.combined.CombinedQueries;
 import io.github.ulisse1996.jaorm.spi.common.Singleton;
 
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.function.Supplier;
+import java.util.stream.Collectors;
+import java.util.stream.StreamSupport;
 
 public abstract class QueriesService {
 
     private static final Singleton<QueriesService> INSTANCE = Singleton.instance();
 
-    public static QueriesService getInstance() {
+    public static synchronized QueriesService getInstance() {
         if (!INSTANCE.isPresent()) {
-            INSTANCE.set(ServiceFinder.loadService(QueriesService.class));
+            List<QueriesService> services = StreamSupport.stream(ServiceFinder.loadServices(QueriesService.class).spliterator(), false)
+                            .collect(Collectors.toList());
+            if (services.size() == 1) {
+                INSTANCE.set(services.get(0));
+            } else {
+                INSTANCE.set(new CombinedQueries(services));
+            }
         }
 
         return INSTANCE.get();
@@ -32,24 +42,26 @@ public abstract class QueriesService {
 
     @SuppressWarnings("unchecked")
     public <T> BaseDao<T> getBaseDao(Class<T> klass) {
+        Class<?> found;
         if (isDelegateClass(klass)) {
-            klass = (Class<T>) DelegatesService.getInstance().getEntityClass(klass);
+            found = DelegatesService.getInstance().getEntityClass(klass);
+        } else {
+            found = klass;
         }
 
-        Class<T> finalKlass = klass;
         return (BaseDao<T>) getQueries()
                 .values()
                 .stream()
-                .filter(entry -> entry.getEntityClass().equals(finalKlass))
+                .filter(entry -> entry.getEntityClass().equals(found))
                 .findFirst()
                 .map(DaoImplementation::getDaoSupplier)
                 .map(Supplier::get)
-                .orElseThrow(() -> new IllegalArgumentException("Can't find BaseDao for " + finalKlass));
+                .orElseThrow(() -> new IllegalArgumentException("Can't find BaseDao for " + found));
     }
 
-    protected <T> boolean isDelegateClass(Class<T> klass) {
+    public <T> boolean isDelegateClass(Class<T> klass) {
         return EntityDelegate.class.isAssignableFrom(klass);
     }
 
-    protected abstract Map<Class<?>, DaoImplementation> getQueries(); //NOSONAR
+    public abstract Map<Class<?>, DaoImplementation> getQueries(); //NOSONAR
 }