+++ jaorm-processor/src/main/java/io/github/ulisse1996/jaorm/processor/JaormProcessor.java
@@ -1,14 +1,12 @@
 package io.github.ulisse1996.jaorm.processor;
 
+import io.github.ulisse1996.jaorm.processor.config.ConfigHolder;
 import io.github.ulisse1996.jaorm.processor.generation.GenerationType;
 import io.github.ulisse1996.jaorm.processor.generation.Generator;
 import io.github.ulisse1996.jaorm.processor.validation.Validator;
 import io.github.ulisse1996.jaorm.processor.validation.ValidatorType;
 
-import javax.annotation.processing.AbstractProcessor;
-import javax.annotation.processing.RoundEnvironment;
-import javax.annotation.processing.SupportedAnnotationTypes;
-import javax.annotation.processing.SupportedSourceVersion;
+import javax.annotation.processing.*;
 import javax.lang.model.SourceVersion;
 import javax.lang.model.element.Element;
 import javax.lang.model.element.TypeElement;
@@ -18,12 +16,15 @@ import java.util.stream.Collectors;
 
 @SupportedAnnotationTypes("io.github.ulisse1996.jaorm.annotation.*")
 @SupportedSourceVersion(SourceVersion.RELEASE_8)
+@SupportedOptions("jaorm.tables.suffix")
 public class JaormProcessor extends AbstractProcessor {
 
     @Override
     public boolean process(Set<? extends TypeElement> annotations, RoundEnvironment roundEnv) {
+        ConfigHolder.init(processingEnv.getOptions());
         validate(roundEnv);
         generate(roundEnv);
+        ConfigHolder.destroy();
         return true;
     }