+++ jaorm-tools/jaorm-maven-plugin/src/main/java/io/github/ulisse1996/jaorm/validation/model/EntityMetadata.java
@@ -8,9 +8,13 @@ import com.github.javaparser.ast.type.Type;
 import com.github.javaparser.resolution.types.ResolvedType;
 import io.github.ulisse1996.jaorm.annotation.Column;
 import io.github.ulisse1996.jaorm.annotation.Converter;
+import io.github.ulisse1996.jaorm.entity.EntityDelegate;
+import io.github.ulisse1996.jaorm.entity.EntityMapper;
 import io.github.ulisse1996.jaorm.validation.util.ValidationUtils;
 
 import java.util.*;
+import java.util.function.Supplier;
+import java.util.stream.Collectors;
 
 public class EntityMetadata {
 
@@ -29,6 +33,7 @@ public class EntityMetadata {
     }
 
     private final List<FieldMetadata> fields;
+    private final String table;
 
     public EntityMetadata(ClassOrInterfaceDeclaration klass) {
         List<FieldMetadata> values = new ArrayList<>();
@@ -39,10 +44,28 @@ public class EntityMetadata {
             }
         }
         this.fields = Collections.unmodifiableList(values);
+        this.table = null;
+    }
+
+    public EntityMetadata(Map.Entry<Class<?>, Supplier<? extends EntityDelegate<?>>> entry) {
+        EntityDelegate<?> delegate = entry.getValue().get();
+        this.fields = buildFields(delegate.getEntityMapper());
+        this.table = delegate.getTable();
+    }
+
+    private List<FieldMetadata> buildFields(EntityMapper<?> entityMapper) {
+        return entityMapper.getMappers()
+                .stream()
+                .map(m -> new FieldMetadata(null, m.getType().getName(), m.getName()))
+                .collect(Collectors.toList());
     }
 
     public List<FieldMetadata> getFields() {
         return fields;
+    }
+
+    public String getTable() {
+        return this.table;
     }
 
     public static class FieldMetadata {
@@ -51,6 +74,12 @@ public class EntityMetadata {
         private final String type;
         private final String columnName;
         private String converterType;
+
+        private FieldMetadata(String name, String type, String columnName) {
+            this.name = name;
+            this.type = type;
+            this.columnName = columnName;
+        }
 
         private FieldMetadata(FieldDeclaration field) {
             VariableDeclarator variableDeclarator = field.getVariables().get(0);
@@ -89,7 +118,7 @@ public class EntityMetadata {
         }
 
         public String getName() {
-            return name;
+            return Optional.ofNullable(this.name).orElse(this.columnName);
         }
 
         public String getType() {