+++ jaorm-tools/jaorm-maven-plugin/src/main/java/io/github/ulisse1996/jaorm/validation/model/enums/ValidationType.java
@@ -1,0 +1,7 @@
+package io.github.ulisse1996.jaorm.validation.model.enums;
+
+public enum ValidationType {
+    CLASSPATH,
+    SOURCE,
+    ALL
+}